Re-add `HandlerType`, `MessageType` and related collection types.
